Cleaned warnings

+++ client/src/ResourceStack.cpp
@@ -96,10 +96,10 @@ namespace zappy
 	    float s = m_cube.scale().y;
 	    float c = (static_cast<float>(m_count) - 1.0f) * g + s;
 
-	    m_cube.scale(1.0, c / s, 1.0);
-	    m_cube.setPosition(m_position + glm::vec3(0, c / 2.0f - 0.01, 0));
+	    m_cube.scale(1.0, static_cast<double>(c / s), 1.0);
+	    m_cube.setPosition(m_position + glm::vec3(0, static_cast<double>(c) / 2.0 - 0.01, 0));
 	    win.draw(camera, m_cube);
-	    m_cube.scale(1.0, 1.0 / c * s, 1.0);
+	    m_cube.scale(1.0, 1.0 / static_cast<double>(c * s), 1.0);
 	  }
 	else
 	  {
